oh-tab-6l9: dedupe e2e pollUntil + host binding docs
