+++ tests/e2e/suite/messaging.ts
@@ -1,18 +1,5 @@
 import * as vscode from 'vscode';
-
-// Helper to poll until condition is met
-async function pollUntil(
-  condition: () => Promise<boolean>,
-  timeoutMs: number = 10000,
-  intervalMs: number = 200
-): Promise<void> {
-  const deadline = Date.now() + timeoutMs;
-  while (Date.now() < deadline) {
-    if (await condition()) return;
-    await new Promise((r) => setTimeout(r, intervalMs));
-  }
-  throw new Error(`pollUntil timed out after ${timeoutMs}ms`);
-}
+import { pollUntil } from './pollUntil';
 
 export async function run(): Promise<void> {
   // Ensure chat view is created
